pyproject.toml
[project]
name = "workflow-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "mistralai>=1.9.2",
    "mistralai-workflows[agents]==2.0.0b5",
    "mistralai-workflows[mistralai]",
    "httpx>=0.27,<1.0",
]
